Start moving the SSE terms into their own file and reduce the MRIS dependency slightly
