+++ utils/mrisurf_integrate.cpp
@@ -17,7 +17,7 @@
  *
  */
 #include "mrisurf_integrate.h"
-
+#include "mrisurf_sseTerms.h"
 
 static int mrisIntegrationEpoch     (MRI_SURFACE *mris, INTEGRATION_PARMS *parms, int n_avgs);
 static double mrisLineMinimize      (MRI_SURFACE *mris, INTEGRATION_PARMS *parms);
@@ -620,7 +620,7 @@ int MRISintegrate(MRI_SURFACE *mris, INTEGRATION_PARMS *parms, int n_averages)
         if (v->ripflag) {
           continue;
         }
-        sq = mrisSampleMinimizationEnergy(mris, v, parms, v->x, v->y, v->z);
+        sq = mrisSampleMinimizationEnergy(mris, vno, parms, v->x, v->y, v->z);
         mn += sqrt(sq);
       }
       mn /= (double)MRISvalidVertices(mris);
